% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-ao.R
\name{download_ao}
\alias{download_ao}
\title{Download North Atlantic Oscillation data}
\usage{
download_ao()
}
\value{
\itemize{
\item Month: Month of record
\item Year: Year of record
\item NAO: North Atlantic Oscillation
}
}
\description{
surface sea-level pressure difference between the Subtropical (Azores) High and the Subpolar Low.
}
\examples{
\dontrun{
nao <- download_nao()
}

}
\references{
\url{https://www.ncdc.noaa.gov/teleconnections/nao}
}
