% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-npgo.R
\name{download_npgo}
\alias{download_npgo}
\title{Download North Pacific Gyre Oscillation data}
\usage{
download_npgo(create_csv = FALSE)
}
\arguments{
\item{create_csv}{Logical option to create a local copy of the data. Defaults to FALSE.}
}
\value{
\itemize{
\item Year
\item Month: Month of record
\item North Pacific Gyre Oscillation
}
}
\description{
North Pacific Gyre Oscillation data
}
\examples{
npgo <- download_npgo()

}
\references{
\url{http://www.o3d.org/npgo/}
}
