% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-pdo.R
\name{download_pdo}
\alias{download_pdo}
\title{Download Pacific Decadal Oscillation Data}
\usage{
download_pdo(use_cache = FALSE, file = NULL)
}
\arguments{
\item{use_cache}{logical option to save and load from cache. If `TRUE`, results will be cached in memory
if `file` is `NULL` or on disk if `file` is not `NULL`.}

\item{file}{optional character with the full path of a file to save the data. If `cache` is `FALSE` but
`file` is not `NULL`, the results will be downloaded from the internet and saved on disk.}
}
\value{
\itemize{
\item Date: Date object that uses the first of the month as a placeholder. Date formatted as date on the first of the month because R only supports one partial of date time
\item Month: Month of record
\item Year: Year of record
\item PDO: Pacific Decadal Oscillation index
}
}
\description{
The PDO index is derived as the leading PC of monthly SST anomalies in the North Pacific Ocean, poleward of 20N. The monthly mean global average SST anomalies are removed to separate this pattern of variability from any "global warming" signal that may be present in the data.
}
\examples{
\dontrun{
pdo <- download_pdo()
}

}
\references{
\url{http://www.cpc.ncep.noaa.gov/products/analysis_monitoring/ensostuff/detrend.nino34.ascii.txt}
}
