% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-roni.R
\name{download_roni}
\alias{download_roni}
\title{Download Relative Oceanic Nino Index data}
\usage{
download_roni(use_cache = FALSE, file = NULL)
}
\arguments{
\item{use_cache}{logical option to save and load from cache. If `TRUE`, results will be cached in memory
if `file` is `NULL` or on disk if `file` is not `NULL`.}

\item{file}{optional character with the full path of a file to save the data. If `cache` is `FALSE` but
`file` is not `NULL`, the results will be downloaded from the internet and saved on disk.}
}
\value{
\itemize{
\item Year: Year of record
\item Season: Season of record
\item roni: Relative Oceanic Niño Index, using the 1991–2020 base period "3 month running mean of ERSST.v5 SST anomalies in the Niño 3.4 region (5°N–5°S, 120°–170°W) with average tropical mean (20°N–20°S) SST anomalies subtracted. The difference is then adjusted so the variance equals the original Niño 3.4 index".
\item Start_Month: Start month of record
\item End_Month: End month of record
\item phase: ENSO phase
}
}
\description{
El Niño–Southern Oscillation (ENSO) is often characterized through the use of sea surface temperature (SST) departures from their climatological values, as in the Niño-3.4 index. However, this approach is problematic in a changing climate when the climatology itself is varying. To address this issue, van Oldenborgh et al. proposed a relative Niño-3.4 SST index, which subtracts the tropical mean SST anomaly from the Niño-3.4 index and multiplies by a scaling factor.
\itemize{
\item The cold phase was defined as periods in which the RONI values within a sliding five-season window were all below −0.5 degC
\item The warm phase was defined as periods in which the RONI values within a sliding five-season window were all above 0.5 degC
\item The neutral phase was defined as the situation outside the definitions of warm phase and cold phase
}
}
\examples{
\dontrun{
roni <- download_roni()
}

}
\references{
\url{https://www.cpc.ncep.noaa.gov/products/analysis_monitoring/enso/roni/#latest-data}
\url{https://journals.ametsoc.org/view/journals/clim/37/4/JCLI-D-23-0406.1.xml}
}
