% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-soi.R
\name{download_soi}
\alias{download_soi}
\title{Download Southern Oscillation index data}
\usage{
download_soi()
}
\description{
Temperature anomalies:
\itemize{
\item Warm phase of El Nino/ Southern Oscillation when 3-month average sea-surface temperature departure of positive 0.5 degC
\item Cool phase of La Nina/ Southern Oscillation when 3-month average sea-surface temperature departure of negative 0.5 degC
\item Neutral phase is defined as the three month temperature average is between +0.5 and -0.5 degC
}
}
\examples{
soi_df <- download_soi()

}

