% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ok_NL}
\alias{ok_NL}
\title{Reference monthly climatology grid 1990-2017 for Temperature, the Netherlands}
\format{An object of class \code{RasterBrick} of dimension 334 x 286 x 336.}
\usage{
data(ok_NL)
}
\description{
raster values interpolated with ordinary kriging, based on daily values.
}
\examples{
data(ok_NL)
summary(ok_NL)
}
\author{
Marieke Dirksen
}
\keyword{Netherlands}
\keyword{datasets}
\keyword{grid}
\keyword{raster}
\keyword{reference}
