% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RR_obs}
\alias{RR_obs}
\title{Monthly precipitation observations for Indonesia}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 840 rows and 1096 columns.}
\usage{
data(RR_obs)
}
\description{
This dataset includes monthly precipitation sums for Indonesia since 1930 until 2000. Coordinates are stored
as data attributes, with "x" as LON and "y" as LAT in WGS84.
}
\examples{
data(RR_obs)
head(RR_obs)
head(attr(RR_obs,'x'))
head(attr(RR_obs,'y'))
}
\author{
Marieke Dirksen
}
\keyword{Indonesia}
\keyword{datasets}
\keyword{observation}
\keyword{precipitation}
