% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RR_ref_grid}
\alias{RR_ref_grid}
\title{Reference monthly climatology grid 1980-2010 for Precipitation, Indonesia}
\format{An object of class \code{RasterBrick} of dimension 100 x 248 x 360.}
\usage{
data(RR_ref_grid)
}
\description{
raster stack object
}
\examples{
data(RR_ref_grid)
summary(RR_ref_grid)
}
\author{
Marieke Dirksen
}
\keyword{Indonesia}
\keyword{datasets}
\keyword{grid}
