% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsoi-package.R
\docType{package}
\name{rsoi-package}
\alias{rsoi}
\alias{rsoi-package}
\title{El Nino/Southern Oscillation (ENSO) and Related Climate Indices}
\description{
Downloads Southern Oscillation Index, Oceanic Nino Index, North Pacific Gyre Oscillation data, North Atlantic Oscillation and Arctic Oscillation. Data sources are described in the help files for each function.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/boshek/rsoi/}
  \item \url{https://boshek.github.io/rsoi/}
}

}
\author{
\strong{Maintainer}: Sam Albers \email{sam.albers@gmail.com}

Other contributors:
\itemize{
  \item Elio Campitelli \email{elio.campitelli@cima.fcen.uba.ar} [contributor]
}

}
\keyword{internal}
