% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsoi-package.R
\docType{package}
\name{rsoi}
\alias{rsoi}
\alias{rsoi-package}
\alias{rsoi-package}
\title{\code{rsoi} package}
\description{
El Nino/Southern Oscillation (ENSO) and Related Climate Indices
}
\details{
See the README on
\href{https://CRAN.R-project.org/package=rsoi}{CRAN}
or \href{https://github.com/boshek/rsoi/blob/master/README.md}{GitHub}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/boshek/rsoi}
  \item Report bugs at \url{https://github.com/boshek/rsoi/issues}
}

}
\author{
\strong{Maintainer}: Sam Albers \email{sam.albers@gmail.com}

}
\keyword{internal}
