% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soi_df}
\alias{soi_df}
\title{SOI index.}
\format{A data frame of SOI
\itemize{
\item Date: Date object that uses the first of the month as a placeholder
\item Month: Month of record
\item Year: Year of record
\item SOI: Southern Oscillation Index
\item SOI_3MON_AVG: 3 Month Average Southern Oscillation Index
\item phase: ENSO phase  
}}
\source{
\url{https://www.ncdc.noaa.gov/teleconnections/enso/indicators/soi/data.csv}
}
\usage{
soi_df
}
\description{
SOI index.
}
\keyword{datasets}

