% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp.RR_stations}
\alias{sp.RR_stations}
\title{Stations with Monthly precipitation observations for Indonesia}
\format{An object of class \code{SpatialPointsDataFrame} with 1095 rows and 4 columns.}
\usage{
data(sp.RR_stations)
}
\description{
This dataset includes monthly precipitation stations for Indonesia since 1930 until 2000.
}
\examples{
library(mapview)
data("sp.RR_stations")
mapview(sp.RR_stations)
}
\author{
Marieke Dirksen
}
\keyword{Indonesia}
\keyword{datasets}
\keyword{observation}
\keyword{precipitation}
\keyword{stations}
