% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_grid_nl}
\alias{sp_grid_nl}
\title{Specification of the grid parameters with time dimension}
\format{An object of class \code{array} of length 1.}
\usage{
data(sp_grid_nl)
}
\description{
Coordinates are stored as data attributes, with "x" as LON and "y" as LAT in WGS84. The time dimension is in months.
}
\examples{
data(sp_grid_nl)
head(attr(sp_grid_nl,'x'))
head(attr(sp_grid_nl,'y'))
head(attr(sp_grid_nl,'time'))
}
\author{
Marieke Dirksen
}
\keyword{Netherlands}
\keyword{datasets}
\keyword{grid}
