% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TN_observations}
\alias{TN_observations}
\title{Monthly temperature observations from the Netherlands}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 1366 rows and 37 columns.}
\usage{
data(TN_observations)
}
\description{
This dataset includes monthly mean temperature observations from the Netherlands since 1906. Coordinates are stored
as data attributes, with "x" as LON and "y" as LAT in WGS84.
}
\examples{
data(TN_observations)
head(TN_observations)
head(attr(TN_observations,'x'))
head(attr(TN_observations,'y'))
}
\author{
Marieke Dirksen
}
\keyword{datasets}
