% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TN_reference}
\alias{TN_reference}
\title{Reference monthly temperature observations from the Netherlands 1990-2017}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 336 rows and 36 columns.}
\usage{
data(TN_reference)
}
\description{
This dataset includes monthly mean temperature observations from the Netherlands from 1990 until 2017. Coordinates are stored
as data attributes, with "x" as LON and "y" as LAT in WGS84.
}
\examples{
data(TN_reference)
head(TN_reference)
head(attr(TN_reference,'x'))
head(attr(TN_reference,'y'))
}
\author{
Marieke Dirksen
}
\keyword{Netherlands}
\keyword{datasets}
\keyword{reference}
\keyword{temperature}
