% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-asymsam.R
\name{download_asymsam_monthly}
\alias{download_asymsam_monthly}
\alias{download_asymsam_daily}
\title{Download Asymmetric and Symmetric SAM indices}
\usage{
download_asymsam_monthly(use_cache = FALSE, file = NULL)

download_asymsam_daily(levels = 700, use_cache = FALSE, file = NULL)
}
\arguments{
\item{use_cache}{logical option to save and load from cache. If `TRUE`, results will be cached in memory
if `file` is `NULL` or on disk if `file` is not `NULL`.}

\item{file}{optional character with the full path of a file to save the data. If `cache` is `FALSE` but
`file` is not `NULL`, the results will be downloaded from the internet and saved on disk.}

\item{level}{atmospheric levels in hPa to download.}
}
\value{
\itemize{
\item Lev: Atmospheric level in hPa
\item Date: Date object that uses the first of the month as a placeholder. Date formatted as date on the first of the month because R only supports one partial of date time
\item Index: Type of index. Either "sam", "ssam" or "asam".
\item Value: Value of the index
\item Value_normalized: Value of the index normalized by the standard deviation of the index
\item R.squared: The variance explained by the index (only in the daily version)
}
}
\description{
The Asymmetric and Symmetric SAM indices are computed as the 
projection of geopotential height anomalies onto the zonally asymmetric and 
zonally symmetric parts of the SAM field.
}
\examples{
\dontrun{
asymsam <- download_asymsam_monthly()
}

}
\references{
\url{https://www.cima.fcen.uba.ar/~elio.campitelli/asymsam/}
}
