% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-npgo.R
\name{download_npgo}
\alias{download_npgo}
\title{Download North Pacific Gyre Oscillation data}
\usage{
download_npgo()
}
\value{
\itemize{
\item Date
\item Year
\item Month: Month of record
\item NPGO: North Pacific Gyre Oscillation
}
}
\description{
North Pacific Gyre Oscillation data
}
\examples{
npgo <- download_npgo()

}
\references{
\url{http://www.o3d.org/npgo/}
}
