% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-soi.R
\name{download_soi}
\alias{download_soi}
\title{Download Southern Oscillation index data}
\usage{
download_soi(create_csv = FALSE)
}
\arguments{
\item{create_csv}{Logical option to create a local copy of the data. Defaults to FALSE.}
}
\value{
\itemize{
\item Date: Date object that uses the first of the month as a placeholder. YrMon formatted as date on the first of the month because R only supports one partial of date time
\item Month: Month of record
\item Year: Year of record
\item SOI: Southern Oscillation Index
\item SOI_3MON_AVG: 3 Month Average Southern Oscillation Index
\item phase: ENSO phase  
}
}
\description{
Temperature anomalies:
\itemize{
\item Warm phase of El Nino/ Southern Oscillation when 3-month average sea-surface temperature departure of positive 0.5 degC
\item Cool phase of La Nina/ Southern Oscillation when 3-month average sea-surface temperature departure of negative 0.5 degC
\item Neutral phase is defined as the three month temperature average is between +0.5 and -0.5 degC
}
}
\examples{
soi_df <- download_soi()
plot(x = soi_df$Date, y = soi_df$SOI, type = "l")

}
