% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ok_NL}
\alias{ok_NL}
\title{Reference climatology grid 1990-2017}
\format{An object of class \code{RasterLayer} of dimension 316 x 267 x 1.}
\usage{
data(ok_NL)
}
\description{
raster values interpolated with ordinary kriging, based on daily values.
}
\examples{
data(ok_NL)
summary(ok_NL)
}
\author{
Marieke Dirksen
}
\keyword{Netherlands}
\keyword{datasets}
\keyword{grid}
\keyword{raster}
\keyword{reference}
