% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsoi-package.R
\docType{package}
\name{rsoi-package}
\alias{rsoi}
\alias{rsoi-package}
\title{El Nino/Southern Oscillation (ENSO) and Related Climate Indices}
\description{
See the README on
\href{https://CRAN.R-project.org/package=rsoi}{CRAN}
or \href{https://github.com/boshek/rsoi/blob/master/README.md}{GitHub}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/boshek/rsoi}
}

}
\author{
\strong{Maintainer}: Sam Albers \email{sam.albers@gmail.com} (0000-0002-9270-7884)

}
\keyword{internal}
