% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsoi.R
\docType{package}
\name{rsoi}
\alias{rsoi}
\alias{rsoi-package}
\title{\code{rsoi} package}
\description{
El Nino/Southern Oscillation (ENSO) and Related Climate Indices
}
\details{
See the README on
\href{https://cran.r-project.org/web/packages/rsoi/README.html}{CRAN}
or \href{https://github.com/boshek/rsoi/blob/master/README.md}{GitHub}
}
